Use same translation object everywhere
