+++ src/main/java/no/ndla/taxonomy/rest/v1/ResourceTypeTranslations.java
@@ -15,6 +15,7 @@ import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import no.ndla.taxonomy.domain.ResourceType;
 import no.ndla.taxonomy.domain.exceptions.NotFoundException;
 import no.ndla.taxonomy.repositories.ResourceTypeRepository;
+import no.ndla.taxonomy.service.dtos.TranslationDTO;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.transaction.annotation.Transactional;
@@ -41,10 +42,10 @@ public class ResourceTypeTranslations {
     @GetMapping
     @Operation(summary = "Gets all relevanceTranslations for a single resource type")
     @Transactional(readOnly = true)
-    public List<ResourceTypeTranslationDTO> getAllResourceTypeTranslations(@PathVariable("id") URI id) {
+    public List<TranslationDTO> getAllResourceTypeTranslations(@PathVariable("id") URI id) {
         ResourceType resourceType = resourceTypeRepository.getByPublicId(id);
-        List<ResourceTypeTranslationDTO> result = new ArrayList<>();
-        resourceType.getTranslations().forEach(t -> result.add(new ResourceTypeTranslationDTO() {
+        List<TranslationDTO> result = new ArrayList<>();
+        resourceType.getTranslations().forEach(t -> result.add(new TranslationDTO() {
             {
                 name = t.getName();
                 language = t.getLanguageCode();
@@ -56,13 +57,13 @@ public class ResourceTypeTranslations {
     @GetMapping("/{language}")
     @Operation(summary = "Gets a single translation for a single resource type")
     @Transactional(readOnly = true)
-    public ResourceTypeTranslationDTO getResourceTypeTranslation(@PathVariable("id") URI id,
+    public TranslationDTO getResourceTypeTranslation(@PathVariable("id") URI id,
             @Parameter(description = "ISO-639-1 language code", example = "nb", required = true) @PathVariable("language") String language) {
         ResourceType resourceType = resourceTypeRepository.getByPublicId(id);
         var translation = resourceType.getTranslation(language).orElseThrow(
                 () -> new NotFoundException("translation with language code " + language + " for resource type", id));
 
-        return new ResourceTypeTranslationDTO() {
+        return new TranslationDTO() {
             {
                 name = translation.getName();
                 language = translation.getLanguageCode();
@@ -96,17 +97,6 @@ public class ResourceTypeTranslations {
             resourceType.removeTranslation(language);
             entityManager.persist(resourceType);
         });
-    }
-
-    @Schema(name = "ResourceTypeTranslation")
-    public static class ResourceTypeTranslationDTO {
-        @JsonProperty
-        @Schema(description = "The translated name of the resource type", example = "Article")
-        public String name;
-
-        @JsonProperty
-        @Schema(description = "ISO 639-1 language code", example = "en")
-        public String language;
     }
 
     public static class ResourceTypeTranslationPUT {
